ci: add security audit workflow for Rust dependencies (#53)
This pull request introduces a new GitHub Actions workflow to perform a
security audit on Rust dependencies. The workflow is triggered by
changes to `Cargo.toml` or `Cargo.lock` files.

### New GitHub Actions Workflow:

*
[`.github/workflows/audit.yml`](diffhunk://#diff-93a89f16f6717d1e4e4e27ec20ef916e3355c7aa890ca74cd9b12db30d8a899fR1-R14):
Added a `Security audit` workflow that runs on `ubuntu-latest`. It uses
the `rustsec/audit-check` action to check for vulnerabilities in Rust
dependencies whenever `Cargo.toml` or `Cargo.lock` files are modified.

Signed-off-by: Tim Tran <trandangtrithanh2000@gmail.com>