fix: apply suggestions (#119)
This pull request makes two minor changes to the `nicknamer` project:
one improves code efficiency by removing an unnecessary `.clone()` call,
and the other adds an `aria-hidden` attribute to an SVG element in the
HTML template to improve accessibility.

Code efficiency improvement:

*
[`nicknamer/server/src/name.rs`](diffhunk://#diff-e5ae5478e1c569b28d41f131b339a2991cd19bd8ce91a22847dc5e0ad5bfbc44L288-R288):
Removed the `.clone()` call when creating a `NamesTableTemplate`
instance, as the `names` variable is no longer needed elsewhere.

Accessibility enhancement:

*
[`nicknamer/server/templates/names/names_table.html`](diffhunk://#diff-328edb0e758794231a55bc4a62e42bc1444789ab7883bebb7797c53138f57548R4):
Added the `aria-hidden="true"` attribute to an SVG element to ensure it
is properly hidden from assistive technologies.