ci: add Docker Hub login step to CI workflow (#186)
This pull request adds a Docker Hub login step to the GitHub Actions
workflow for running tests. This change enables the workflow to
authenticate with Docker Hub, which is necessary for pulling private
images or performing other Docker operations that require
authentication.

Continuous integration workflow improvements:

*
[`.github/workflows/run-tests.yml`](diffhunk://#diff-7314d0ebbd2e9537ae4889316745b4fd2fa43cb86275c9caae18a86ba228b642R20-R24):
Added a step to log in to Docker Hub using the `docker/login-action@v3`,
with credentials sourced from GitHub variables and secrets.